requirements.txt
requests==2.31.0
ipython==8.14.0
openai==0.27.8
langchain>=0.3.8,<0.4.0
langchain-community==0.3.8
python-dotenv
PyYAML
opencv-contrib-python
fastapi
uvicorn
pydantic
pydantic[email]
tortoise-orm
Pillow
python-multipart
aiofiles
aiomysql
asyncmy
